requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.65.0
opencv-python==4.8.0.74
pyinstaller==6.0.0
tkinter 